titanic/requirements.txt
django===3.1.0
djangorestframework==3.11.1
tensorflow==2.3.0
numpy==1.18.5
django-cors-headers==3.4.0
mysqlclient==2.0.1
python-dotenv==0.14.0